restructuring- resolving comments from JYB's review
