+++ app.py
@@ -4,15 +4,15 @@ from chalicelib.s3 import S3Client
 from chalicelib.utils import get_file_extension_from_base64
 
 # Blueprints
-from chalicelib.blueprints.announcements import announcements_routes
 
 import uuid
 
 from chalicelib.api.listings import listings_api
 from chalicelib.api.applicants import applicants_api
+from chalicelib.api.announcements import announcements_api
 
 app = Chalice(app_name="zap")
-app.register_blueprint(announcements_routes)
+app.register_blueprint(announcements_api)
 app.register_blueprint(listings_api)
 app.register_blueprint(applicants_api)
